project structure refactor (#45)
* ci: ci update

* ci: ci update

* ci: ci update

* ci: fix

* refactor: project structure refactor

* refactor: project structure refactor
